fix(LIVE-19667): move custom swap screen to swap subScreens
